Fixed upgrade modifier executor input value

+++ src/main/java/mc/rellox/spawnerlegacyapi/modifier/executor/IExecutor.java
@@ -16,7 +16,7 @@ public interface IExecutor {
 	
 	interface IExecutorUpgrades extends IExecutor {
 		
-		void execute(int[] values);
+		void execute(double[] values);
 		
 	}
